+++ jscomp/syntax/src/res_printer.ml
@@ -2416,6 +2416,25 @@ and print_pattern ~state (p : Parsetree.pattern) cmt_tbl =
       Doc.concat [Doc.text "..."; print_ident_path ident cmt_tbl]
     | Ppat_type ident ->
       Doc.concat [Doc.text "#..."; print_ident_path ident cmt_tbl]
+    | Ppat_record (rows, _)
+      when ParsetreeViewer.has_dict_pattern_attribute p.ppat_attributes ->
+      Doc.concat
+        [
+          Doc.text "dict{";
+          Doc.indent
+            (Doc.concat
+               [
+                 Doc.soft_line;
+                 Doc.join
+                   ~sep:(Doc.concat [Doc.text ","; Doc.line])
+                   (List.map
+                      (fun row -> print_pattern_dict_row ~state row cmt_tbl)
+                      rows);
+               ]);
+          Doc.if_breaks (Doc.text ",") Doc.nil;
+          Doc.soft_line;
+          Doc.rbrace;
+        ]
     | Ppat_record (rows, open_flag) ->
       Doc.group
         (Doc.concat
@@ -2581,6 +2600,35 @@ and print_pattern_record_row ~state row cmt_tbl =
            ])
     in
     print_comments doc cmt_tbl loc_for_comments
+
+and print_pattern_dict_row ~state
+    ((longident, pattern) : Longident.t Location.loc * Parsetree.pattern)
+    cmt_tbl =
+  let loc_for_comments =
+    {longident.loc with loc_end = pattern.ppat_loc.loc_end}
+  in
+  let rhs_doc =
+    let doc = print_pattern ~state pattern cmt_tbl in
+    let doc =
+      if Parens.pattern_record_row_rhs pattern then add_parens doc else doc
+    in
+    Doc.concat [print_optional_label pattern.ppat_attributes; doc]
+  in
+  let lbl_doc =
+    Doc.concat [Doc.text "\""; print_longident longident.txt; Doc.text "\""]
+  in
+  let doc =
+    Doc.group
+      (Doc.concat
+         [
+           lbl_doc;
+           Doc.text ":";
+           (if ParsetreeViewer.is_huggable_pattern pattern then
+              Doc.concat [Doc.space; rhs_doc]
+            else Doc.indent (Doc.concat [Doc.line; rhs_doc]));
+         ])
+  in
+  print_comments doc cmt_tbl loc_for_comments
 
 and print_expression_with_comments ~state expr cmt_tbl : Doc.t =
   let doc = print_expression ~state expr cmt_tbl in